GVT-740: Refactor folder structure for location track and reference line panels
