+++ ui/src/map/settings/map-layer-settings.scss
@@ -31,7 +31,7 @@
     cursor: pointer;
 
     &:not(:last-child) {
-        margin-bottom: 12px;
+        margin-bottom: 14px;
     }
 
     &--indented {
