+++ ui/src/selection-panel/location-track-panel/location-track.scss
