+++ ui/src/selection-panel/location-track-panel/location-tracks-panel.tsx
