+++ ui/src/selection-panel/reference-line-panel/reference-line.scss
