+++ ui/src/selection-panel/reference-line-panel/reference-lines-panel.tsx
